Re-split ps1 to match psp
